moved export out of route, another attempt at deployment fix
