+++ app/api/logout/route.tsx
@@ -6,10 +6,10 @@ import { NextResponse } from "next/server";
 import { auth } from "@/lib/firebase-admin";
 import prisma from "@/lib/prisma";
 
+import { AUTH_VALIDITY_IN_DAYS } from "../constants";
 import { getDateNow } from "../login/date";
 
 const dynamic = "force-dynamic";
-export const AUTH_VALIDITY_IN_DAYS = 1;
 
 const checkUserExists = async (email: string): Promise<boolean> => {
   return (
